Prevent the getAllowedValuesForKey function from throwing errors if no allowedValues are specified
Since `undefined` and `null` are not iterable, this function throws an error when `allowedValues` is not specified in the schema. This function, however, is called regardless of the existence of this field by Meteor Autoform. This causes various types of fields to not render. This simple fix should prevent errors without changing the functionality.