Fix #307 - Test for empty object when creating schema
Issue #307 is caused because if you try to import a file with no
exports in ESM, it returns an empty object. This is probably to achieve
interop with CJS, which initializes `module.exports` to an empty object.
This empty object passed validation checks because empty objects are
truthy, and no specific check exists.

To solve this, I've added a specific checks for empty objects in
`checkAndScrubDefinition` (I tried using lodash's isObject, but that
returned true for all type constructors, which also return true for
isEmpty 🤷‍♂️).

From the docs it doesn't look like a plain object is ever a valid type,
so I don't believe this should break anything for anyone else